09-travel-itinerary-agent/requirements.txt


# Apify API client for external data
apify-client>=1.8.0

# Additional dependencies for production
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.32.0
google-adk
google-cloud-aiplatform
cloudpickle
pydantic
pytest
overrides
scikit-learn
reportlab
google-cloud-storage
deprecated
python-multipart
# Optional: For enhanced functionality
# fastapi>=0.115.0  # For web interface
# uvicorn>=0.34.0   # For web server
# streamlit>=1.28.0 # For alternative UI
